Support resolving configuration in output
Also support disabling comments if desired.

At one point there was some problem with calling resolve on the configuration before rendering.

But that was several versions of the hocon library ago, and I haven't been able to find a case where
the output after calling resolve produces unexpected output. And my archaelogy hasn't been able to
discover the exact scenario that caused me to decide not to do the resolution before rendering.

Just in case it does cause issues, I have made it configurable, so resolving references can be
disabled if needed.